# Conflicts:
#	lib/jxl/dec_cache.cc
#	lib/jxl/dec_cache.h
#	lib/jxl/dec_frame.cc
#	lib/jxl/dec_frame.h
#	lib/jxl/dec_group.cc
#	lib/jxl/dec_group.h
#	lib/jxl/dec_modular.cc
#	lib/jxl/dec_reconstruct.cc
#	lib/jxl/epf.cc
#	lib/jxl/filters.cc
#	lib/jxl/filters.h
#	lib/jxl/render_pipeline/low_memory_render_pipeline.cc
#	lib/jxl/render_pipeline/low_memory_render_pipeline.h
#	lib/jxl/render_pipeline/render_pipeline.h
#	lib/jxl/render_pipeline/render_pipeline_stage.h
#	lib/jxl/render_pipeline/stage_blending.cc
#	lib/jxl/render_pipeline/stage_gaborish.cc
#	third_party/lcms
#	tools/epf_main.cc
#	tools/epf_run.cc
#	tools/ssimulacra.cpp
#	tools/ssimulacra.txt
#	tools/ssimulacra_opencv_main.cpp

+++ lib/jxl/render_pipeline/stage_gaborish.cc
@@ -62,8 +62,8 @@ class GaborishStage : public RenderPipelineStage {
 #endif
       // Since GetInputRow(input_rows, c, {-1, 0, 1}) is aligned, rounding
       // xextra up to Lanes(d) doesn't access anything problematic.
-      for (int64_t x = -RoundUpTo(xextra, Lanes(d));
-           x < (int64_t)(xsize + xextra); x += Lanes(d)) {
+      for (ssize_t x = -RoundUpTo(xextra, Lanes(d));
+           x < (ssize_t)(xsize + xextra); x += Lanes(d)) {
         const auto t = LoadMaybeU(d, row_t + x);
         const auto tl = LoadU(d, row_t + x - 1);
         const auto tr = LoadU(d, row_t + x + 1);
